Should be called EndScreen (#12)
* save progress

* Prefab created

makes the fail screen now float on the game screen with transparency. Failing screen has two buttons, restart, which brings to the start of the game, and quit.

* deleted the scene FailState

Added FailState to prefabs, unneccesary to have another scene for it

* added a WinScreen and updated script

the float screen script now allows us to select a state we want a screen to be floating at, and the game object we want to be floated. (don't have to code for every single instance)

* Fixed variable names and added folders to organize

* newline