Pooling Objects
Fixed the tiny bug that caused the bullets to not target the players. But I essentially implemented the pooling aspect to the game objects such that it hopefully drops the frame rate less than before because it doesn't continuously instantiate game objects, but each turret has a pool of bullets to use from.